+++ .gitignore
@@ -41,3 +41,4 @@ proto_define/struct_desc/__pycache__/__init__.cpython-313.pyc
 proto_define/__pycache__/proto_user.cpython-313.pyc
 proto_define/__pycache__/__init__.cpython-313.pyc
 build/
+__pycache__/utils.cpython-313.pyc
